Updated Launch file, working towards independent costmap on costmap_conveter
