+++ .bashrc
