+++ src/stratom_sim/models/conveyor/meshes/conveyor.dae
