+++ src/stratom_sim/models/conveyor/meshes/conveyor2.dae
