+++ src/stratom_sim/models/conveyor/model.config
