+++ src/stratom_sim/models/turtlebot3_common/meshes/burger_base.dae
