+++ src/stratom_sim/models/turtlebot3_common/meshes/r200.dae
