+++ src/stratom_sim/models/turtlebot3_common/meshes/tire.dae
