+++ src/stratom_sim/models/turtlebot3_common/meshes/waffle_base.dae
