+++ src/stratom_sim/models/turtlebot3_common/model.config
