+++ src/stratom_sim/models/workcell/meshes/mesh.dae
