+++ src/stratom_sim/models/workcell/meshes/mesh_backup.dae
