+++ src/stratom_sim/models/workcell/model.config
